refactor(dane): R4/R5/R6/M5 z audytu — dlug architektoniczny
R4. CapPositionRepository.producerSelectionOf() siegal wprost do
CapCacheDao, omijajac CapCacheRepository — ta sama tabela miala dwa
wejscia z warstwy repozytoriow. Wstrzykniecie zmienione na
CapCacheRepository, ktore juz mialo gotowa metode getByIds().

R5. CapCacheRepository.getCountryStats()/getOwnedCapsByCountry()
przepuszczaly modele zapytan Room (@ColumnInfo CountryStatRow,
OwnedCapRow) prosto do czterech ViewModeli. Doszly waskie modele
domenowe CountryCapCount/OwnedCapThumbnail (model/binder/), mapowanie
na granicy repozytorium — kontynuacja ACL z domain_acl_room_leak.

R6. CapsRepository.getById(Int) bylo niespojne z capId: Long w calej
reszcie kodu — piec wolajacych robilo .toInt() osobno. Sygnatura
zmieniona na Long, konwersja do Retrofit @Path zostaje w jednym
miejscu (repozytorium). Przy okazji: _collectionChanged jako
SharedFlow(extraBufferCapacity=1) z tryEmit() moglo po cichu zgubic
drugi sygnal pod rzad przy szybkich zmianach (PurchasedViewModel nie
odswiezalby listy) — tryEmit -> emit (suspend), markPurchasedLocally
tez suspend.

M5. WorldMap.countries wystawial mutowalne android.graphics.Path
wprost z pol WorldMapParser (@Singleton, zyje przez caly proces) —
kontrakt "tylko do odczytu" byl niejawny, transformacja in-place po
stronie UI zepsulaby mape trwale. countries jest teraz computed
property zwracajaca kopie (Path(src)); jedyny konsument
(LocationsMapScreen.kt:192) czyta je raz na obiekt mapy przez
remember(state.map), wiec koszt kopiowania ~200 sciezek jest
pomijalny.

Zweryfikowane: BUILD SUCCESSFUL (testDebugUnitTest +
compileDebugAndroidTestKotlin).

Swiadomie nieruszone (decyzja produktowa 2026-08-07): R2/D2/S3 —
model dopuszcza kilka pozycji na ten sam kapsel; duplikaty kapsli nie
beda wspierane, wiec problem u korzenia nie wystapi.

+++ app/src/main/java/pl/sroki/cci/android/data/CollectionVerifier.kt
@@ -55,7 +55,7 @@ class CollectionVerifier @Inject constructor(
     suspend fun verify(capId: Long, onCatalogReached: () -> Unit = {}): CatalogStatus {
         val stored = capCacheRepository.getOne(capId)
         val cap = try {
-            capsRepository.getById(capId.toInt())
+            capsRepository.getById(capId)
         } catch (e: HttpException) {
             if (e.code() == 404) {
                 // 404 to też odpowiedź katalogu: serwer był osiągalny, kapsel z niego zniknął.
